Update fingerprint, allow for user to specify fingerprint (#7)
Update to new default SSL fingerprint for api.conduit.suyash.io (closes #6) and allows for the user of this library to specify a fingerprint of their own if they choose.

Of course, ideally no fingerprint would be needed for verification (and we could just verify by checking the certificate chain against a known root certificate). This change requires downstream changes in other libraries, however, and is being tracked by #5.